ZEN-33260 ApplyDataMap no longer does an order-insensitive comparison before calling setter
